fix: destroy like by other user

+++ app/controllers/posts/likes_controller.rb
@@ -16,8 +16,9 @@ class Posts::LikesController < Posts::ApplicationController
 
   def destroy
     like = PostLike.find(params[:id])
-    like.destroy!
+    return unless like.user_id == current_user.id
 
+    like.destroy!
     respond_to do |format|
       format.html { redirect_to post_url(like.post_id) }
     end
